Fix link to code of conduct
Replaces the original broken link to code of conduct (which was maybe
working when/if the code of conduct file lived in the top-level of the
repo?) with the link provided by the GitHub UI to the code of conduct
tab. An alternate solution might be to link to the
`.github/CODE_OF_CONDUCT.md` path directly instead.